requirements/production.txt
gunicorn==19.7.1
Flask==0.12.2
Flask-RESTful==0.3.6
flask-marshmallow==0.8.0
pytest==3.5.0
factory-boy==2.10.0
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.3.0
pytest-cov==2.6.0
Flask-Testing==0.7.1